judges/requirements.txt
python-dotenv
requests
bs4
pandas